refactored the code to use re-exports
