+++ frontend/src/components/index.js
@@ -1,0 +1,3 @@
+export { Cabinet } from "./Cabinet";
+export { GameSettingsBox } from "./GameSettingsBox";
+export { NavigationBar } from "./NavigationBar";
