+++ frontend/src/hooks/index.js
@@ -1,0 +1,6 @@
+export { useAuthContext } from "./useAuthContext";
+export { useGameSettingsContext } from "./useGameSettingsContext";
+export { useLoadUpGameSettings } from "./useLoadUpGameSetting";
+export { useLogin } from "./useLogin";
+export { useLogout } from "./useLogOut";
+export { useSignup } from "./useSignup";
