+++ frontend/src/pages/Home.js
@@ -1,13 +1,9 @@
-import { useEffect, useState } from "react";
-import { GameSettingsBox } from "../components/GameSettingsBox";
-import { useLoadUpGameSettings } from "../hooks/useLoadUpGameSetting";
-// import { useGameSettingsContext } from "../hooks/useGameSettingsContext";
-import { Button, Drawer } from "flowbite-react";
-import { Cabinet } from "../components/Cabinet";
+import { useEffect } from "react";
+import { useLoadUpGameSettings } from "../hooks";
+import { Cabinet } from "../components";
+
 export const Home = () => {
-  // const { gameSettings } = useGameSettingsContext();
   const { LoadUpGameSettings } = useLoadUpGameSettings();
-  // const [gameSettingsOpen, setGameSettingsOpen] = useState(false);
 
   useEffect(() => {
     const loadingfunction = async () => {
@@ -16,9 +12,6 @@ export const Home = () => {
     loadingfunction();
   }, []);
 
-  // const handleGameSettingsClose = () => {
-  //   setGameSettingsOpen(false);
-  // };
   return (
     <>
       <div className="text-black dark:text-white z-0">
@@ -27,22 +20,6 @@ export const Home = () => {
       <div className=" absolute right-0 top-14  z-10">
         <Cabinet />
       </div>
-
-      {/* <Button
-        onClick={() => {
-          setGameSettingsOpen(!gameSettingsOpen);
-        }}
-      >
-        open game settings
-      </Button>
-      <Drawer
-        open={gameSettingsOpen}
-        onClose={handleGameSettingsClose}
-        position="right"
-      >
-        <Drawer.Header title="Drawer" />
-        <Drawer.Items>{gameSettings && <GameSettingsBox />}</Drawer.Items>
-      </Drawer> */}
     </>
   );
 };
