+++ frontend/src/pages/index.js
@@ -1,0 +1,4 @@
+export { Home } from "./Home";
+export { Login } from "./Login";
+export { PageNotFound } from "./PageNotFound";
+export { Signup } from "./Signup";
